L2Math  * just some method overloading
git-svn-id: http://svn.l2jfree.com/l2jfree@6199 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/serverpackets/L2GameServerPacket.java
@@ -101,7 +101,7 @@ public abstract class L2GameServerPacket extends SendablePacket<L2GameClient>
 		if (Config.PACKET_FINAL)
 			writeQ(value);
 		else
-			writeD((int)L2Math.limit(Integer.MIN_VALUE, value, Integer.MAX_VALUE));
+			writeD(L2Math.limit(Integer.MIN_VALUE, value, Integer.MAX_VALUE));
 	}
 	
 	protected final void writeCompH(int value)
